+++ src/main/java/com/depromeet/infra/config/fcm/FcmConfig.java
@@ -1,4 +1,4 @@
-package com.depromeet.global.config.fcm;
+package com.depromeet.infra.config.fcm;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
